requirements.txt
playwright
python-dotenv
beautifulsoup4
requests